+++ Tests/MapboxDirectionsTests/RouteOptionsTests.swift
@@ -431,6 +431,110 @@ class RouteOptionsTests: XCTestCase {
         let oneWaypointOptionsWithNonNilName = RouteOptions(waypoints: waypoints)
         XCTAssertFalse(oneWaypointOptionsWithNonNilName.urlQueryItems.map(\.name).contains("waypoint_names"))
     }
+
+    func testRefreshingEnabledInitializationWithQueryItems() {
+        let origin = Waypoint(coordinate: LocationCoordinate2D(latitude: 39.15031, longitude: -84.47182))
+        let destination = Waypoint(coordinate: LocationCoordinate2D(latitude: 39.12971, longitude: -84.51638))
+
+        let options = RouteOptions(waypoints: [origin, destination], profileIdentifier: .automobile)
+        options.refreshingEnabled = true
+        XCTAssertFalse(options.urlQueryItems.contains(URLQueryItem(name: "enable_refresh", value: "true")))
+
+        options.profileIdentifier = .init(rawValue: "custom/driving")
+        XCTAssertFalse(options.urlQueryItems.contains(URLQueryItem(name: "enable_refresh", value: "true")))
+
+        options.profileIdentifier = .automobileAvoidingTraffic
+        XCTAssertTrue(options.urlQueryItems.contains(URLQueryItem(name: "enable_refresh", value: "true")))
+
+        options.profileIdentifier = .init(rawValue: "custom/driving-traffic")
+        XCTAssertTrue(options.urlQueryItems.contains(URLQueryItem(name: "enable_refresh", value: "true")))
+    }
+
+    func testRefreshingEnabledUrlQueryItems() {
+        let origin = Waypoint(coordinate: LocationCoordinate2D(latitude: 39.15031, longitude: -84.47182))
+        let destination = Waypoint(coordinate: LocationCoordinate2D(latitude: 39.12971, longitude: -84.51638))
+        let queryItems: [URLQueryItem] = [.init(name: "enable_refresh", value: "true")]
+
+        let options1 = RouteOptions(
+            waypoints: [origin, destination],
+            profileIdentifier: .automobile,
+            queryItems: queryItems
+        )
+        XCTAssertFalse(options1.refreshingEnabled)
+
+        let options2 = RouteOptions(
+            waypoints: [origin, destination],
+            profileIdentifier: .automobileAvoidingTraffic,
+            queryItems: queryItems
+        )
+        XCTAssertTrue(options2.refreshingEnabled)
+
+        let options3 = RouteOptions(
+            waypoints: [origin, destination],
+            profileIdentifier: .init(rawValue: "custom/driving-traffic"),
+            queryItems: queryItems
+        )
+        XCTAssertTrue(options3.refreshingEnabled)
+    }
+
+    func testAllowsUTurnAtWaypoint() {
+        let options1 = RouteOptions(coordinates: testCoordinates, profileIdentifier: .automobile)
+        XCTAssertFalse(options1.allowsUTurnAtWaypoint)
+
+        let options2 = RouteOptions(coordinates: testCoordinates, profileIdentifier: .automobileAvoidingTraffic)
+        XCTAssertFalse(options2.allowsUTurnAtWaypoint)
+
+        let options3 = RouteOptions(
+            coordinates: testCoordinates,
+            profileIdentifier: ProfileIdentifier(rawValue: "custom/driving-traffic")
+        )
+        XCTAssertFalse(options3.allowsUTurnAtWaypoint)
+
+        let options4 = RouteOptions(
+            coordinates: testCoordinates,
+            profileIdentifier: ProfileIdentifier(rawValue: "custom/driving")
+        )
+        XCTAssertFalse(options4.allowsUTurnAtWaypoint)
+
+        let options5 = RouteOptions(coordinates: testCoordinates, profileIdentifier: .walking)
+        XCTAssertTrue(options5.allowsUTurnAtWaypoint)
+
+        let options6 = RouteOptions(coordinates: testCoordinates, profileIdentifier: .cycling)
+        XCTAssertTrue(options6.allowsUTurnAtWaypoint)
+    }
+
+    func testArriveByUrlQueryItems() {
+        let options = RouteOptions(coordinates: testCoordinates, profileIdentifier: .automobile)
+        options.arriveBy = Date(timeIntervalSince1970: 600)
+        XCTAssertTrue(options.urlQueryItems.contains(URLQueryItem(name: "arrive_by", value: "1970-01-01T00:10")))
+
+        options.profileIdentifier = ProfileIdentifier(rawValue: "custom/driving")
+        XCTAssertTrue(options.urlQueryItems.map { $0.name }.contains("arrive_by"))
+
+        options.profileIdentifier = .automobileAvoidingTraffic
+        XCTAssertFalse(options.urlQueryItems.map { $0.name }.contains("arrive_by"))
+
+        options.profileIdentifier = .walking
+        XCTAssertFalse(options.urlQueryItems.map { $0.name }.contains("arrive_by"))
+    }
+
+    func testDepartAtUrlQueryItems() {
+        let options = RouteOptions(coordinates: testCoordinates, profileIdentifier: .automobile)
+        options.departAt = Date(timeIntervalSince1970: 600)
+        XCTAssertTrue(options.urlQueryItems.contains(URLQueryItem(name: "depart_at", value: "1970-01-01T00:10")))
+
+        options.profileIdentifier = ProfileIdentifier(rawValue: "custom/driving")
+        XCTAssertTrue(options.urlQueryItems.map { $0.name }.contains("depart_at"))
+
+        options.profileIdentifier = .automobileAvoidingTraffic
+        XCTAssertTrue(options.urlQueryItems.map { $0.name }.contains("depart_at"))
+
+        options.profileIdentifier = ProfileIdentifier(rawValue: "custom/driving-traffic")
+        XCTAssertTrue(options.urlQueryItems.map { $0.name }.contains("depart_at"))
+
+        options.profileIdentifier = .walking
+        XCTAssertFalse(options.urlQueryItems.map { $0.name }.contains("depart_at"))
+    }
 }
 
 private let testCoordinates = [